feat: stub out the command line interface
This creates the commands for `run`, `serve` and `submit` but only implements `run`.

This introduces a `stepflow-config.yml` that can be used to configure the available plugins.
See `examples/python` for how to set it up.

If there is a config, it is used as the working directory for the subprocesses.

If there is a flow, it is used as the search directory to locate the config.

JSON and YAML files are supported for most things (not the `stepflow-config.yml`
since I haven't yet implemented a richer search). Ideally, that would consider
`stepflow[-_]config.{yaml,yml,json}`. Likely a glob library I could use for that.