feat: Change reference syntax
This came up when looking at syntax for referencing different parts of steps, etc.
The new syntax is more uniform -- it is always a `$ref` and then a struct
with fields indicating what is being referenced.

In the future, we can allow `$ref` to be a string that is parsed into one of these
structs. This allows us to treat the struct as the "reference format", while still
providing an expression-like shorthand for people writing flows directly.