feat: Introduce workflow execution loop
This changes the workflow execution to using a loop that submits
the runnable steps for execution, and adds new steps as those
complete. This structure is more amenable to journalling and
restarts, as well as debugging / single-stepping and other
execution patterns compared to the relatively rigid "create the
topology and run it".