build: Change workflow to use manual dispatch
This was necessary because PRs / tags created by GitHub actions
do not trigger other GitHub actions. This meant that the release
tag (created when the PR merged) didn't trigger the release build.

To fix this, I instead have the PR manually trigger the release build,
which now does the build and *then* creates the release and tag.