feat: Introduce ValueTemplate (#125)
* feat: Introduce ValueTemplate

This is a JSON value that can reference workflow inputs and the results
of earlier steps.

This type simplifies the use of expressions within values, making it
apparent in the type system that (a) expressions can appear, and (b)
allowing us to actually represent the expressions rather than re-parsing
them from a raw JSON value when used.

This also ensures that the structure of these value templates and
expressions appear in the `flow.json`.

* chore: comments; fmt