ci: Setup CI for python (#139)
* ci: Setup CI for python

Changes `ci.yml` to split actual logic into separate actions, and use
the edited paths to control which tests run.

Introduces an integration test CI job which runs on changes to the core
runtime (stepflow-rs) or the Python SDK (sdks/python).

Applies formatting and linting to the Python SDK (typechecking
to-be-fixed in a later PR).

Removes the "run all the tests / examples" form Rust since that was
covering a combination of Rust and Python, and uses the StepFlow binary
in the integration job.

* chroe: licensure

* test: fix test scripts