feat: add eval method and demonstrate loop/map
Also fixed some problems with parsing $from in ValueTemplate which
caused errors to be suppressed / tests to fail.

Since the `flows/evaluate` method allows invoking a flow (and therofore
takes a flow as an argument) the protocol schema now includes the flow
schema. As a result, I removed the separate `flow.json` schema file
and the corresponding generated code since they were overlapping.