feat: add eval method and demonstrate loop/map (#153)
* feat: add eval method and demonstrate loop/map

Also fixed some problems with parsing $from in ValueTemplate which
caused errors to be suppressed / tests to fail.

Since the `flows/evaluate` method allows invoking a flow (and therofore
takes a flow as an argument) the protocol schema now includes the flow
schema. As a result, I removed the separate `flow.json` schema file
and the corresponding generated code since they were overlapping.

* chore: python lint/fmt

* chore: python types

* chore: rust fmt/clippy

* docs: fix docs references to flow.json

* chore: python fmt & type-check

* chore: fix types and schema

Main issue was that the `FlowResult` schema wasn't generating a JSON
schema that encoded the discriminant in a way that the Python code
generator supported.

* chore: fix fmt, docs, etc.