feat: Change to trie-based routing (#157)
This allows routing rules to define the component passed to the plugin,
allowing `/python/{*component}` to be passed as `/{component}` to the
Python plugin, and removes the need to track the "prefix" each plugin
was available at (which would be inconsistent if available via multiple
routes anyway).

This also allows adds the ability to filter components used in a rule.

This fixes the list of components by applying a reverse routing step
to indicate the routes that each component is available at.