feat: replace HTTP+SSE with Streamable HTTP
This implements the newer Streamable HTTP transport described in the
MCP specification. The main benefit is that each request is sent via
POST and only use SSE for bidirectional communication when the component
uses `context` (indicating it may potentially call methods or send
notifications to the client).

Also adds the ability to start and stop test servers in the workflow
test configuration, allowing the HTTP tranpsort to be tested in
/tests/http.