feat: replace HTTP+SSE with Streamable HTTP (#168)
* feat: replace HTTP+SSE with Streamable HTTP

This implements the newer Streamable HTTP transport described in the
MCP specification. The main benefit is that each request is sent via
POST and only use SSE for bidirectional communication when the component
uses `context` (indicating it may potentially call methods or send
notifications to the client).

Also adds the ability to start and stop test servers in the workflow
test configuration, allowing the HTTP tranpsort to be tested in
/tests/http.

* chore: cleanup message decoder

* choer: cleanup server.py

* chore: cleanup stdio_server and tests

* chore: cleanup test_http_server.py

* chore: cleanups in http/client.rs

* chore: use async stream

* chore: fix Python fmt/lint

* chore: minor client logging cleaunps

* chore: more concurrency

* chore: add notify_dyn