fix: Notification responses (#184)
* fix: Notification responses

Python was sending responses for notification requests over stdio.
This was likely introduced with the Streamable HTTP work, since we
do need to send the 202 response (with an empty body) for notifications.

This change removes the empty notification response when using the
stdio transport.

It also removes the unnecessary plugin name (which was incompletely
removed during the routing changes).

* chore: fmt, remove unused script