feat: Use blobs for flows (#195)
This introduces a "blob type" which allows blobs to record whether they
store data, flows, etc.

This replaces the separate flow storage with the typed blob storage.

This changes to pass the flow ID rather than the flow itself when
evaluating a flow. This reduces the need to copy the flow (eg., when
executing it for each item of a list, it only needs to be sent once to
get the blob ID).

This also changes some code to standardize towards the term "flow"
rather than "workflow".

This also introduces a `execute_flow_by_id` method which allows
executing a flow by its ID, rather than needing to fetch the flow first.
This allows sharing logic between the `eval` builtin and the `execute`
method.