feat: Remove fusion and special cases for tools/embeddings (#330)
The reason that we needed to support "complex configuration" was because
of non-serializable values (like Embeddings). However, all of the Embeddings
extend pydantic `BaseModel`, which means they *are* serializable. This change
allows us to serialize embeddings, removing the need for (and code for) fusion
and fused-execution.

This required special handling to ensure SecretStr values (such as
`OPENAI_API_KEY` are serialized). We'll likely want to explore a better
and more secure option (like writing those to a separate secret store
and referencing them in the JSON).

This also changed how tools are made available to agents. Rather than
special casing them, we detect components that are set to `tool_mode`
and create a serializable description of the component, including the
custom code. This allows us to recreate the component during
deserialization and pass it to the agent (or other consumer of the tool).