feat: use variables for load_from_db inputs (#424)
This uses Stepflow variables to implement `load_from_db` inputs.
This allows passing the variable when executing the flow.

This is part of #320 (specifically, part 2, 3 and 4) of the langflow
steps identified https://github.com/stepflow-ai/stepflow/issues/320#issuecomment-3549510640).

This also simplifies some of the python code (removing dead code) and fixes a bug in JsonPath (where `null` wasn't parsing as an empty JsonPath).