refactor: Eliminate duplicate JsonSchema crate (#460)
* refactor: Eliminate duplicate JsonSchema crate

This eliminates JsonSchema in favor of using utoipa everywhere.

This required some additional configurations / code since utoipa
doesn't support dumping the JsonSchema (it is focused on OpenAPI).

It was easier to dump the JsonSchema (which utoipa could provide the
information for) than it would be to integrate JsonSchema with the
API services.

This closes #440.

* chore: lint, review comments

* build: fix serde_with dependency

* fix: ensure proper generation of flow result